refactor: centralize proficiency data
